refactoring에 따른 오류 수정
