+++ frontend/assj/src/component/kakaomap/modal.css
@@ -3,8 +3,10 @@
   display: none;
   position: fixed;
   z-index: 999;
-  width: 100%;
-  height: 100%;
+  top: 0;
+  right: 0;
+  bottom: 0;
+  left: 0;
   background-color: rgba(0, 0, 0, 0.6);
   font-family: "Montserrat", sans-serif;
 }
